css/css-position/sticky-pos.html

<title>'sticky' element's top percentage value should be resolved against proper box</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3115" />
<meta name="assert" content="This test checks that for which container
sticky inset percentage should be resolved." />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>
<div style="height: 500px; background: purple; overflow: hidden;">
  <div style="height: 400px; background: yellow;">
    <div id="target" style="height: 100px; position: -webkit-sticky; position: sticky; left: 0; top: 50%; background: blue;">
    </div>
  </div>
</div>
<script>
test(() => {
  assert_equals(getComputedStyle(target).top, '250px');
}, "Sticky element's top property percentage value should be resolved against the div with overflow: hidden");
</script>
